requirements.txt
streamlit
torch
torchvision
pillow
gdown
numpy
pyttsx3
